+++ src/main/kotlin/io/kjson/util/ArrayUtil.kt
@@ -1,8 +1,8 @@
 /*
- * @(#) ResourceLoaderTest.kt
+ * @(#) ArrayUtil.kt
  *
  * resource-loader  Resource loading mechanism
- * Copyright (c) 2021, 2024, 2025 Peter Wall
+ * Copyright (c) 2025 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,34 +23,14 @@
  * SOFTWARE.
  */
 
-package io.kjson.resource
+package io.kjson.util
 
-import kotlin.test.Test
-import kotlin.test.fail
-
-import java.io.File
-import java.net.URL
-
-import io.kstuff.test.shouldBe
-import io.kstuff.test.shouldBeOneOf
-import io.kstuff.test.shouldEndWith
-
-class ResourceTest {
-
-    @Test fun `should display readable form of URL on toString`() {
-        val resource1 = XMLLoader().resource(File("src/test/resources/xml/test.xml"))
-        resource1.toString() shouldBe "src/test/resources/xml/test.xml"
-        val resource2 = XMLLoader().resource(URL("http://kjson.io/xml/test9.xml"))
-        resource2.toString() shouldBe "http://kjson.io/xml/test9.xml"
-    }
-
-    @Test fun `should get a classpath URL`() {
-        val url = Resource.classPathURL("/xml/test2.xml") ?: fail("Can't locate resource")
-        url.protocol shouldBeOneOf listOf("file", "jar")
-        url.toString() shouldEndWith "/xml/test2.xml"
-        val resource = XMLLoader().resource(url)
-        val document = resource.load()
-        document.documentElement.tagName shouldBe "test2"
-    }
-
+/**
+ * Return `true` if this array starts with the given sequence of elements.
+ */
+fun <X> Array<X>.startsWith(other: Array<X>): Boolean {
+    for (i in other.indices)
+        if (this[i] != other[i])
+            return false
+    return true
 }